+++ build_script.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
 rm -f *.out 
-g++ -g -Wall -Wextra -pedantic main.cpp src/helper.cpp -o json_parser.out 
+g++ -g -Wall -Wextra -pedantic main.cpp src/parser.cpp -o json_parser.out
